+++ applications/services/gui/view_holder.c
