+++ applications/services/gui/view_holder.h
